Patch 6: horizgrid
- Added patch "horizgrid"
- fixed comments in config.def.h
================================
Additional Information:

horizontal grid
Description

This patch is a variant of gaplessgrid. It arranges windows in a grid pattern in which every window is roughly the same size, adjusted such that there are no gaps. However, this layout arranges the windows in a horizontal grid, rather than a vertical grid.
Horizontal Grid Layout

horizgrid     (###)
+--------+--------+
|        |        |
|        |        |
+-----+--+--+-----+
|     |     |     |
|     |     |     |
+-----+-----+-----+

gapless_grid
+--------+--------+
|        |        |
|        +--------+
+--------+        |
|        +--------+
|        |        |
+--------+--------+